Refactoring & reconfigure classes.
