--- HomeWork_9/Shapes.java
@@ -1,19 +1,0 @@
-package HomeWork_9;
-
-import java.util.Scanner;
-
-import static java.awt.Color.blue;
-import static java.awt.Color.getColor;
-
-public class Shapes {
-
-    protected String  color;
-
-    protected void calculateArea() {
-
-    }
-    protected void calculatePerimeter() {
-
-    }
-}
-
