+++ HomeWork_9/Shapes/Main_Shapes.java
@@ -1,4 +1,4 @@
-package HomeWork_9;
+package HomeWork_9.Shapes;
 
 public class Main_Shapes {
     public static void main(String[] args) {
